// src/enums/searchType.ts
export enum SearchType {
    EQUAL = 'equal',
    EQUAL_IN_PART = 'equal_in_part',
    RANGE = 'range',
    SEARCH = 'search',
    RANGE_DATE = 'rangeDate',
    SORT_ORDER = 'sortOrder',
    KEYWORD = 'keyword',
    SORT_BY_LOCATION = 'sortByLocation',
    SELECT = 'select',
    PERCENTAGE_RANGE = 'percentage_range',
    USD_RANGE = 'usd_range',
    CHECKED_RANGE = 'checked_range',
}

export enum SortOrder {
    ASC = 'ASC',
    DESC = 'DESC',
}

export enum KeyCode {
    ENTER = 'Enter',
    ARROW_DOWN = 'ArrowDown',
    ARROW_UP = 'ArrowUp',
}