// src/enums/cellNameType.ts
export enum CellNameType {
    CHECKBOX = 'checkbox',
    NAME = 'name',
    FOUNDED_AT = 'founded_at',
    COUNTRY_CODE = 'country_code',
    REGION = 'region',
    CITY = 'city',
    DESCRIPTION = 'description',
    INDUSTRIES = 'industries',
    LAST_FUNDING_STAGE = 'last_funding_stage',
    TOTAL_FUNDING = 'total_funding',
    LAST_FUNDING_DATE = 'last_funding_date',
    PROBABILITY_OF_SUCCESS = 'probability_of_success',
    FRACTION_FEMALE_FOUNDERS = 'fraction_female_founders',
    FRACTION_MISSING_GENDER_FOUNDERS = 'fraction_missing_gender_founders',
    FRACTION_MALE_FOUNDERS = 'fraction_male_founders',
    FRACTION_OTHER_FOUNDERS = 'fraction_other_founders',
    FRACTION_FEMALE_EXECUTIVES = 'fraction_female_executives',
    FRACTION_MALE_EXECUTIVES = 'fraction_male_executives',
    FRACTION_OTHER_EXECUTIVES = 'fraction_other_executives',
    FRACTION_MISSING_GENDER_EXECUTIVES = 'fraction_missing_gender_executives',
    FOUNDER_IS_CEO = 'founder_is_ceo',
    BUSINESS_ROLES_COUNT = 'business_roles_count',
    OTHER_ROLES_COUNT = 'other_roles_count',
    FRACTION_BUSINESS_ROLES = 'fraction_business_roles',
    EXECUTIVES_COUNT = 'executives_count',
    FOUNDERS_COUNT = 'founders_count',
    TEAM_FOUNDERS_IN_EXECUTIVES = 'team_founders_in_executives',
    INSTITUTION_NAME_NUNIQUE_FOUNDERS = 'institution_name_nunique_founders',
    INSTITUTION_NAME_COUNT_FOUNDERS = 'institution_name_count_founders',
    INSTITUTION_NAME_COUNT_EXECUTIVES = 'institution_name_count_executives',
    SUBJECT_AGGREGATED_NUNIQUE_FOUNDERS = 'subject_aggregated_nunique_founders',
    DEGREE_TYPE_NUMERIC_MEAN_FOUNDERS = 'degree_type_numeric_mean_founders',
    DEGREE_TYPE_NUMERIC_MAX_FOUNDERS = 'degree_type_numeric_max_founders',
    DEGREE_TYPE_AGGREGATED_NUNIQUE_EXECUTIVES = 'degree_type_aggregated_nunique_executives',
    DEGREE_TYPE_NUMERIC_STD_FOUNDERS = 'degree_type_numeric_std_founders',
    DEGREE_TYPE_AGGREGATED_NUNIQUE_FOUNDERS = 'degree_type_aggregated_nunique_founders',
    SUBJECT_AGGREGATED_NUNIQUE_EXECUTIVES = 'subject_aggregated_nunique_executives',
    DEGREE_TYPE_NUMERIC_MEAN_EXECUTIVES = 'degree_type_numeric_mean_executives',
    DEGREE_TYPE_NUMERIC_STD_EXECUTIVES = 'degree_type_numeric_std_executives',
    DEGREE_TYPE_NUMERIC_MAX_EXECUTIVES = 'degree_type_numeric_max_executives',
    UPDATED_AT = 'updated_at',
}